# Introduce `UPSERT` in `set_trial_user_attr`