--- a/pyproject.toml
+++ b/pyproject.toml
@@ -35,7 +35,7 @@ dependencies = [
   "colorlog",
   "numpy",
   "packaging>=20.0",
-  "sqlalchemy>=1.3.0",
+  "sqlalchemy>=1.4.0",
   "tqdm",
   "PyYAML",  # Only used in `optuna/cli.py`.
 ]