@nabenabe0928 2024-10-22
@porink0424 

> [!NOTE]
> As in the last comment, this PR does not speed up for MySQL.