@c-bata 2024-10-22
> @porink0424
> 
> Note
> 
> As in the last comment, this PR does not speed up for MySQL.

@nabenabe0928 It looks that your benchmark results shows this PR speeds up `trial.set_user_attr()` calls ( `master` 4955.39ms → `this branch` 4420.31ms). Is that correct?