@nabenabe0928 2024-10-22
```python
sqlite_pr = [5725.75, 5891.18, 5837.73, 5904.36, 5686.25, 5548.71, 5537.34, 5380.70, 5567.75, 5666.13]
mysql_pr = [4411.56, 4386.89, 4507.84, 4388.80, 4434.83, 4484.78, 4477.91, 4637.17, 4672.88, 4323.81]

sqlite_master = [5801.52, 5888.66, 7761.65, 6197.31, 6175.02, 6288.58, 6482.06, 5978.97, 5893.87, 5895.68]
mysql_master = [4588.37, 4585.37, 4594.20, 5047.10, 4829.42, 4913.89, 4271.38, 4795.67, 4378.23, 4953.27]
```

@c-bata @porink0424 
I checked the runtime again over 10 times and then I could see the speedup, sorry for the confusion :pray: