@not522 2024-10-23
In my local environment (Mac), the speed has not improved and is equivalent to master. But introducing this PR has the benefit of reducing the number of queries, so I think it is good to merge.
I am checking the range of environments (SQLAlchemy / MySQL / SQLite / ...) in which this PR works.